+++ packages/traceloop-sdk/traceloop/sdk/tracing/tracing.py
@@ -3,6 +3,7 @@ import logging
 import os
 import importlib.util
 
+
 from colorama import Fore
 from opentelemetry import trace
 from opentelemetry.exporter.otlp.proto.http.trace_exporter import (
@@ -25,9 +26,10 @@ from opentelemetry.context import get_value, attach, set_value
 
 from opentelemetry.semconv.ai import SpanAttributes
 from traceloop.sdk import Telemetry
+from traceloop.sdk.instruments import Instruments
 from traceloop.sdk.tracing.content_allow_list import ContentAllowList
 from traceloop.sdk.utils import is_notebook
-from typing import Dict
+from typing import Dict, Optional, Set
 
 TRACER_NAME = "traceloop.tracer"
 EXCLUDED_URLS = """
@@ -55,6 +57,7 @@ class TracerWrapper(object):
         processor: SpanProcessor = None,
         propagator: TextMapPropagator = None,
         exporter: SpanExporter = None,
+        instruments: Optional[Set[Instruments]] = None,
     ) -> "TracerWrapper":
         if not hasattr(cls, "instance"):
             obj = cls.instance = super(TracerWrapper, cls).__new__(cls)
@@ -108,7 +111,134 @@ class TracerWrapper(object):
             if propagator:
                 set_global_textmap(propagator)
 
-            init_instrumentations()
+            instrument_set = False
+            if instruments is None:
+                init_instrumentations()
+                instrument_set = True
+            else:
+                for instrument in instruments:
+                    if instrument == Instruments.OPENAI:
+                        if not init_openai_instrumentor():
+                            print(Fore.RED + "Warning: OpenAI library does not exist.")
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.ANTHROPIC:
+                        if not init_anthropic_instrumentor():
+                            print(
+                                Fore.RED + "Warning: Anthropic library does not exist."
+                            )
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.COHERE:
+                        if not init_cohere_instrumentor():
+                            print(Fore.RED + "Warning: Cohere library does not exist.")
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.PINECONE:
+                        if not init_pinecone_instrumentor():
+                            print(
+                                Fore.RED + "Warning: Pinecone library does not exist."
+                            )
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.CHROMA:
+                        if not init_chroma_instrumentor():
+                            print(Fore.RED + "Warning: Chroma library does not exist.")
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.LANGCHAIN:
+                        if not init_langchain_instrumentor():
+                            print(
+                                Fore.RED + "Warning: LangChain library does not exist."
+                            )
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.LLAMA_INDEX:
+                        if not init_llama_index_instrumentor():
+                            print(
+                                Fore.RED + "Warning: LlamaIndex library does not exist."
+                            )
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.TRANSFORMERS:
+                        if not init_transformers_instrumentor():
+                            print(
+                                Fore.RED
+                                + "Warning: Transformers library does not exist."
+                            )
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.REQUESTS:
+                        if not init_requests_instrumentor():
+                            print(
+                                Fore.RED + "Warning: Requests library does not exist."
+                            )
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.URLLIB3:
+                        if not init_urllib3_instrumentor():
+                            print(Fore.RED + "Warning: urllib3 library does not exist.")
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.PYMYSQL:
+                        if not init_pymysql_instrumentor():
+                            print(Fore.RED + "Warning: PyMySQL library does not exist.")
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.BEDROCK:
+                        if not init_bedrock_instrumentor():
+                            print(Fore.RED + "Warning: Bedrock library does not exist.")
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.REPLICATE:
+                        if not init_replicate_instrumentor():
+                            print(
+                                Fore.RED + "Warning: Replicate library does not exist."
+                            )
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.VERTEXAI:
+                        if not init_vertexai_instrumentor():
+                            print(
+                                Fore.RED + "Warning: Vertex AI library does not exist."
+                            )
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    elif instrument == Instruments.WATSONX:
+                        if not init_watsonx_instrumentor():
+                            print(Fore.RED + "Warning: Watsonx library does not exist.")
+                            print(Fore.RESET)
+                        else:
+                            instrument_set = True
+                    else:
+                        print(
+                            Fore.RED
+                            + "Warning: "
+                            + instrument
+                            + " library does not exist."
+                        )
+                        print(Fore.RESET)
+
+            if not instrument_set:
+                print(
+                    Fore.RED + "Warning: No valid instruments set. Remove 'instrument' "
+                    "argument to use all instruments, or set a valid instrument."
+                )
+                print(Fore.RESET)
 
             obj.__content_allow_list = ContentAllowList()
 
@@ -261,7 +391,7 @@ def init_instrumentations():
     init_cohere_instrumentor()
     init_pinecone_instrumentor()
     init_chroma_instrumentor()
-    # init_haystack_instrumentor()
+    init_haystack_instrumentor()
     init_langchain_instrumentor()
     init_llama_index_instrumentor()
     init_transformers_instrumentor()
@@ -271,6 +401,7 @@ def init_instrumentations():
     init_bedrock_instrumentor()
     init_replicate_instrumentor()
     init_vertexai_instrumentor()
+    init_watsonx_instrumentor()
 
 
 def init_openai_instrumentor():
@@ -281,6 +412,7 @@ def init_openai_instrumentor():
         instrumentor = OpenAIInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_anthropic_instrumentor():
@@ -291,6 +423,7 @@ def init_anthropic_instrumentor():
         instrumentor = AnthropicInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_cohere_instrumentor():
@@ -301,6 +434,7 @@ def init_cohere_instrumentor():
         instrumentor = CohereInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_pinecone_instrumentor():
@@ -311,6 +445,7 @@ def init_pinecone_instrumentor():
         instrumentor = PineconeInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_chroma_instrumentor():
@@ -321,6 +456,7 @@ def init_chroma_instrumentor():
         instrumentor = ChromaInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_haystack_instrumentor():
@@ -331,6 +467,7 @@ def init_haystack_instrumentor():
         instrumentor = HaystackInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_langchain_instrumentor():
@@ -341,6 +478,7 @@ def init_langchain_instrumentor():
         instrumentor = LangchainInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_transformers_instrumentor():
@@ -351,6 +489,7 @@ def init_transformers_instrumentor():
         instrumentor = TransformersInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_llama_index_instrumentor():
@@ -361,6 +500,7 @@ def init_llama_index_instrumentor():
         instrumentor = LlamaIndexInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_requests_instrumentor():
@@ -370,6 +510,7 @@ def init_requests_instrumentor():
         instrumentor = RequestsInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument(excluded_urls=EXCLUDED_URLS)
+    return True
 
 
 def init_urllib3_instrumentor():
@@ -379,6 +520,7 @@ def init_urllib3_instrumentor():
         instrumentor = URLLib3Instrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument(excluded_urls=EXCLUDED_URLS)
+    return True
 
 
 def init_pymysql_instrumentor():
@@ -388,6 +530,7 @@ def init_pymysql_instrumentor():
         instrumentor = PyMySQLInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_bedrock_instrumentor():
@@ -397,6 +540,7 @@ def init_bedrock_instrumentor():
         instrumentor = BedrockInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_replicate_instrumentor():
@@ -407,6 +551,7 @@ def init_replicate_instrumentor():
         instrumentor = ReplicateInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
 
 
 def init_vertexai_instrumentor():
@@ -417,3 +562,15 @@ def init_vertexai_instrumentor():
         instrumentor = VertexAIInstrumentor()
         if not instrumentor.is_instrumented_by_opentelemetry:
             instrumentor.instrument()
+    return True
+
+
+def init_watsonx_instrumentor():
+    if importlib.util.find_spec("ibm_watson_machine_learning") is not None:
+        Telemetry().capture("instrumentation:watsonx:init")
+        from opentelemetry.instrumentation.watsonx import WatsonxInstrumentor
+
+        instrumentor = WatsonxInstrumentor()
+        if not instrumentor.is_instrumented_by_opentelemetry:
+            instrumentor.instrument()
+    return True